Encapsulate unsafe code in a raw module
